toolkit/components/extensions/test/mochitest/test_ext_webrequest_hsts.html

<!DOCTYPE HTML>

<html>
<head>
<meta charset="utf-8">
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <script type="text/javascript" src="head_webrequest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<script>
"use strict";

function getExtension() {
  async function background() {
    let expect;
    let urls = ["*://*.example.org/tests/*"];
    browser.webRequest.onBeforeRequest.addListener(details => {
      browser.test.assertEq(expect.shift(), "onBeforeRequest");
    }, {urls}, ["blocking"]);
    browser.webRequest.onBeforeSendHeaders.addListener(details => {
      browser.test.assertEq(expect.shift(), "onBeforeSendHeaders");
    }, {urls}, ["blocking", "requestHeaders"]);
    browser.webRequest.onSendHeaders.addListener(details => {
      browser.test.assertEq(expect.shift(), "onSendHeaders");
    }, {urls}, ["requestHeaders"]);
    browser.webRequest.onHeadersReceived.addListener(details => {
      browser.test.assertEq(expect.shift(), "onHeadersReceived");

      let headers = details.responseHeaders || [];
      for (let header of headers) {
        if (header.name.toLowerCase() === "strict-transport-security") {
          return;
        }
      }

      headers.push({
        name: "Strict-Transport-Security",
        value: "max-age=31536000000",
      });
      return {responseHeaders: headers};
    }, {urls}, ["blocking", "responseHeaders"]);
    browser.webRequest.onBeforeRedirect.addListener(details => {
      browser.test.assertEq(expect.shift(), "onBeforeRedirect");
    }, {urls});
    browser.webRequest.onResponseStarted.addListener(details => {
      browser.test.assertEq(expect.shift(), "onResponseStarted");
    }, {urls});
    browser.webRequest.onCompleted.addListener(details => {
      browser.test.assertEq(expect.shift(), "onCompleted");
      browser.test.sendMessage("onCompleted", details.url);
    }, {urls});
    browser.webRequest.onErrorOccurred.addListener(details => {
      browser.test.notifyFail(`onErrorOccurred ${JSON.stringify(details)}`);
    }, {urls});

    async function onUpdated(tabId, tabInfo, tab) {
      if (tabInfo.status !== "complete") {
        return;
      }
      browser.tabs.remove(tabId);
      browser.tabs.onUpdated.removeListener(onUpdated);
      browser.test.sendMessage("tabs-done", tab.url);
    }
    browser.test.onMessage.addListener((url, expected) => {
      expect = expected;
      browser.tabs.onUpdated.addListener(onUpdated);
      browser.tabs.create({url});
    });
  }

  let manifest = {
    "permissions": [
      "tabs",
      "webRequest",
      "webRequestBlocking",
      "<all_urls>",
    ],
  };
  return ExtensionTestUtils.loadExtension({
    manifest,
    background,
  });
}

// This test makes a request against a server that redirects with a 302.
add_task(async function test_hsts_request() {
  const testPath = "example.org/tests/toolkit/components/extensions/test/mochitest";

  let extension = getExtension();
  await extension.startup();

  // simple redirect
  let sample = "https://example.org/tests/toolkit/components/extensions/test/mochitest/file_sample.html";
  extension.sendMessage(
    `https://${testPath}/redirect_auto.sjs?redirect_uri=${sample}`,
    ["onBeforeRequest", "onBeforeSendHeaders", "onSendHeaders",
     "onHeadersReceived", "onBeforeRedirect", "onBeforeRequest",
     "onBeforeSendHeaders", "onSendHeaders", "onHeadersReceived",
     "onResponseStarted", "onCompleted"]);
  // redirect_auto adds a query string
  ok((await extension.awaitMessage("tabs-done")).startsWith(sample), "redirection ok");
  ok((await extension.awaitMessage("onCompleted")).startsWith(sample), "redirection ok");

  // priming hsts
  extension.sendMessage(
    `https://${testPath}/hsts.sjs`,
    ["onBeforeRequest", "onBeforeSendHeaders", "onSendHeaders",
     "onHeadersReceived", "onResponseStarted", "onCompleted"]);
  is(await extension.awaitMessage("tabs-done"),
     "https://example.org/tests/toolkit/components/extensions/test/mochitest/hsts.sjs",
     "hsts primed");
  is(await extension.awaitMessage("onCompleted"),
     "https://example.org/tests/toolkit/components/extensions/test/mochitest/hsts.sjs");

  // test upgrade
  extension.sendMessage(
    `http://${testPath}/hsts.sjs`,
    ["onBeforeRequest", "onBeforeRedirect", "onBeforeRequest",
     "onBeforeSendHeaders", "onSendHeaders", "onHeadersReceived",
     "onResponseStarted", "onCompleted"]);
  is(await extension.awaitMessage("tabs-done"),
     "https://example.org/tests/toolkit/components/extensions/test/mochitest/hsts.sjs",
     "hsts upgraded");
  is(await extension.awaitMessage("onCompleted"),
     "https://example.org/tests/toolkit/components/extensions/test/mochitest/hsts.sjs");

  await extension.unload();
});

// This test makes a priming request and adds the STS header, then tests the upgrade.
add_task(async function test_hsts_header() {
  const testPath = "test1.example.org/tests/toolkit/components/extensions/test/mochitest";

  let extension = getExtension();
  await extension.startup();

  // priming hsts, this time there is no STS header, onHeadersReceived adds it.
  let completed = extension.awaitMessage("onCompleted");
  let tabdone = extension.awaitMessage("tabs-done");
  extension.sendMessage(
    `https://${testPath}/file_sample.html`,
    ["onBeforeRequest", "onBeforeSendHeaders", "onSendHeaders",
     "onHeadersReceived", "onResponseStarted", "onCompleted"]);
  is(await tabdone, `https://${testPath}/file_sample.html`, "priming request done");
  is(await completed, `https://${testPath}/file_sample.html`, "priming request done");

  // test upgrade from http to https due to onHeadersReceived adding STS header
  completed = extension.awaitMessage("onCompleted");
  tabdone = extension.awaitMessage("tabs-done");
  extension.sendMessage(
    `http://${testPath}/file_sample.html`,
    ["onBeforeRequest", "onBeforeRedirect", "onBeforeRequest",
     "onBeforeSendHeaders", "onSendHeaders", "onHeadersReceived",
     "onResponseStarted", "onCompleted"]);
  is(await tabdone, `https://${testPath}/file_sample.html`, "hsts upgraded");
  is(await completed, `https://${testPath}/file_sample.html`, "request upgraded");

  await extension.unload();
});
</script>
</head>
<body>

</body>
</html>
